Day8: ViewModels + Add Product Form (With Ids) [INCOMPLETE]

+++ src/main/java/com/Ecomercial/CTTT2018/controllers/ProductController.java
@@ -28,7 +28,7 @@ public class ProductController {
 
     //@PreAuthorize("hasAuthority('ADMIN')")
     @RequestMapping(value = "/product/view/{id}", method = RequestMethod.GET)
-    public ModelAndView viewProduct(@PathVariable("id") long id) {
+    public ModelAndView viewProduct(@PathVariable("id") Integer id) {
 
         logger.info("Product Controller: show product view page(get)");
         Optional<Product> product = productService.getProductById(id);
